Devide code to more logical packages && refactor any namespaces
